Module 2 – SDLC Assignment
Managed to get it working, but the 3rd most common word is a " " at 35 entries, which I cannot figure out how to circumvent. Instruction on setting file directory for the FileReader class to work is located near the top of the class. Also, didn't realize until after starting on the project that a FileReader class already exists. Whoops.